Update Gasto.py
Corrige error dtypes de actualizacion

+++ pages/Gasto.py
@@ -406,10 +406,13 @@ filtered_data = data0[
 ultimo_mes_real = filtered_data['Período'].max()
 gasto_acumulado_real = filtered_data[filtered_data['Período'] <= ultimo_mes_real]['Valor/mon.inf.'].sum()
 
+# Asegurarse de que 'Mes' en budget_data es de tipo entero
+budget_data['Mes'] = budget_data['Mes'].astype(int)
+
 # Calcular el gasto acumulado presupuestado usando el DataFrame de presupuesto filtrado
 filtered_budget_data = budget_data[
     (budget_data['Año'].isin(selected_years)) & 
-    (budget_data['Mes'].astype(int) <= ultimo_mes_real)
+    (budget_data['Mes'] <= ultimo_mes_real)
 ]
 gasto_acumulado_presupuestado = filtered_budget_data['Presupuesto'].sum() if not filtered_budget_data.empty else None
